workers/zap/requirements.txt
celery
python-owasp-zap-v2.4
requests
